48HSK/RAG/python-rag/requirements.txt
openai>=1.50.0
python-dotenv==1.0.0
numpy>=1.26.0
pandas>=2.2.0
sentence-transformers
groq